example: media: watch: do not deinit current wf
Previously, pausing the clock application set all watch face to the deinit state.

This was not ideal. This change prevents pausing the clock application from
setting the current watch face to the deinit state.

This will prevent re-entering the wf init function when the clock application is
re-entered. Instead, the system will resume from the resume state.

This commit will works if this issue (#182) is fixed.

Signed-off-by: Troy Mitchell <troymitchell988@gmail.com>